requirements.txt
google-generativeai>=0.3.0
nltk>=3.8
tenacity>=8.2.0
python-dotenv>=1.0.0
tqdm>=4.66.0
